Let's try again...

+++ trunk/includes/nodetree.php
@@ -235,7 +235,9 @@ class CMS_Node {
 		}
 		$cache->destroy('sql', NODES_TABLE);
 		$cache->destroy('sql', NODE_REVISIONS_TABLE);
-		$cache->destroy('_page_output');
+		if (defined('IN_ADMIN')) {
+			$cache->destroy('_page_output');
+		}
 		// and now add/update the revision :)
 		if ($all) {
 			utils::get_types();
